requirements.txt
apt-clone
apturl
beautifulsoup4
Brlapi
certifi
chardet
colorama
command-not-found
configobj
cupshelpers
dbus-python
defer
distro
eyeD3
filelock
Flask-Cors
GitPython
grpcio
gyp
idna
ifaddr
IMDbPY
importlib-metadata
jeepney
launchpadlib
louis
macaroonbakery
Mako
more-itertools
nemo-emblems
netaddr
netifaces
oauthlib
onboard
packaging
PAM
paramiko
pexpect
Pillow
pip
protobuf
psutil
ptyprocess
pycups
pycurl
pyelftools
pygit2
PyGithub
PyGObject
PyICU
pyinotify
pymacaroons
pyparted
pyRFC3339
python-apt
python-debian
python-gnupg
python-Levenshtein
python-magic
python-xlib
pytz
pyxdg
PyYAML
reportlab
requests-file
screen-resolution-extra
SecretStorage
setproctitle
setuptools
soupsieve
swiftly-sys
systemd-python
tinycss2
tldextract
ubuntu-drivers-common
ufw
ulauncher
Unidecode
urllib3
wadllib
websocket-client
wheel
xdg
xkit
xlrd
youtube-dl
zipp
